refactor: mongodb document 또한 스키마 관리하도록 및 응답 통일
